+++ NEWS.md
@@ -1,3 +1,4 @@
 # comprep (development version)
 
 * Create function to compare recalculated with reported value
+* Create function to compare recalculated p-values with reported p-values